modified:   AUTHORITY/myCA.pem 	modified:   AUTHORITY/recreate_CA_certificate.sh 	modified:   BAK/CA_configuration.cfg 	renamed:    COMMON/create_own_key.sh -> BAK/OLD.create_own_key.sh 	new file:   BAK/OLD.create_own_pem.sh 	renamed:    REDHAT/OWN_extensions.cfg -> BAK/OWN_all-purpose_extensions.cfg 	renamed:    REDHAT/OWN_org_details.cfg -> BAK/OWN_org_details.cfg 	renamed:    COMMON/create_new_CA_key.sh -> BAK/create_new_CA_key.sh 	renamed:    COMMON/create_new_CA_pem.sh -> BAK/create_new_CA_pem.sh 	renamed:    COMMON/use_provided_CA.sh -> BAK/use_provided_CA.sh 	new file:   BAK/vars.sh 	renamed:    REDHAT/CA_configuration.cfg -> COMMON/CA_configuration.cfg 	new file:   COMMON/OWN_all-purpose_extensions.cfg 	new file:   COMMON/OWN_org_details.cfg 	deleted:    COMMON/create_own_pem.sh 	modified:   COMMON/start_over.sh 	new file:   COMMON/step_1.sh 	new file:   COMMON/step_2.sh 	new file:   COMMON/step_3.sh 	new file:   COMMON/step_4.sh 	deleted:    COMMON/validate_CA_pem.sh 	deleted:    COMMON/validate_own_pem.sh 	modified:   COMMON/vars.sh 	new file:   COMMON/view_CA_pem.sh 	new file:   COMMON/view_own_pem.sh 	new file:   DEBIAN/add_CA_to_host_trust_store.sh 	new file:   DEBIAN/functions.sh 	new file:   DEBIAN/test_required_software.sh 	modified:   REDHAT/README.md 	modified:   REDHAT/add_CA_to_host_trust_store.sh 	deleted:    REDHAT/create_new_CA_key.sh 	deleted:    REDHAT/create_new_CA_pem.sh 	deleted:    REDHAT/create_own_key.sh 	deleted:    REDHAT/create_own_pem.sh 	new file:   REDHAT/step_1.sh 	new file:   REDHAT/step_2.sh 	new file:   REDHAT/step_3.sh 	new file:   REDHAT/step_4.sh 	modified:   REDHAT/test_required_software.sh 	deleted:    REDHAT/use_provided_CA.sh 	deleted:    REDHAT/validate_CA_pem.sh         deleted:    REDHAT/validate_own_pem.sh 	deleted:    REDHAT/vars.sh 	new file:   REDHAT/view_CA_pem.sh 	new file:   REDHAT/view_own_pem.sh

+++ BAK/create_new_CA_key.sh
@@ -3,7 +3,15 @@
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
 cd $SCRIPT_DIR
 unset ERR
-source ${SCRIPT_DIR}/functions.sh  
+source ${SCRIPT_DIR}/functions.sh 
+source ${SCRIPT_DIR}/vars.sh
+
+if [ "$(basename $SCRIPT_DIR)" == "COMMON" ]; then
+  TEMP_COLOR=lred; print_color "Do not run scripts from the COMMON directory. "; unset TEMP_COLOR; print_color "Exiting.\n\n"; exit 1;
+fi
+
+
+if [ ! "${RSA}" == "4096" ]; then RSA=2048; fi
 
 # if the file exists but it is empty, remove it
 if [ ! -s myCA.key ] && [ -f myCA.key ]; then rm -f myCA.key; fi
@@ -16,7 +24,7 @@ fi
 TEMP_COLOR=lcyan; print_color "You must enter a passphrase for your key two times. "
 TEMP_COLOR=lred; print_color "Do not lose your passphrase!\n"; unset TEMP_COLOR
 
-openssl genrsa -des3 -out myCA.key 2048 || ERR=TRUE
+openssl genrsa -des3 -out myCA.key ${RSA} || ERR=TRUE
 
 if [ $ERR ]; then 
   TEMP_COLOR=lred; print_color "Something went wrong. Perhaps the passphrases did not match. Try again.\n"; unset TEMP_COLOR;
